+++ autd3-driver/src/datagram/stm/mod.rs
@@ -1,5 +1,6 @@
 mod foci;
 mod gain;
+mod sampling_config;
 
 pub use foci::FociSTM;
 pub use gain::GainSTM;