+++ autd3/src/prelude.rs
@@ -23,8 +23,8 @@ pub use autd3_driver::{
     firmware::{
         cpu::GainSTMMode,
         fpga::{
-            DebugType, Drive, EmitIntensity, GPIOIn, GPIOOut, LoopBehavior, Phase, STMConfig,
-            SamplingConfig, Segment, TransitionMode,
+            DebugType, Drive, EmitIntensity, GPIOIn, GPIOOut, LoopBehavior, Phase, SamplingConfig,
+            Segment, TransitionMode,
         },
         operation::SilencerTarget,
     },