+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - `Controller::link` and `geometry` is now private. use `link()`, `link_mut()`, `geometry()`, and `geometry_mut()` instead
 - Unified `FociSTM` and `GainSTM` constructor to `new`
 - Remove `Silencer::from_completion_steps`, use `Silencer::from_completion_time` instead
+- Change the maximum value of `LoopBehavior` to 65535 from 4294967295
+- Change base frequency of `SamplingConfig` to 40kHz from 20.48MHz
 - Change to use `NonZero<T>` for integer argment which should be non-zero
 - Change `Simulator::builder` argument from port number to `SocketAddr`
 - Make `gain::Group` parallelism controllable